* fixed a couple of warnings - isolated sigslot.h in src directory. There is another one in libs\cocos2dx\extensions\CocoStudio\Armature\external_tool\

+++ dev/project/src/LeaderboardScene.cpp
@@ -129,7 +129,7 @@ void LeaderboardScene::_initWithRecords(UIScrollView* base, UIWidget* entry, con
 
 	// get top of the container and substract entry height as it's anchor is left bottom corner
 	CCPoint position = ccp(0, containerHeight - entryHeight);
-	for (int a = 0; a < records.size(); ++a)
+	for (size_t  a = 0; a < records.size(); ++a)
 	{
 		const LeaderboardEntry& record = records[a];
